#67: [Infrastructure] support@mapzone.io email account (http://github.com/Mapzone/mapzone/issues/67)
Pending changes

+++ io.mapzone.arena/src/io/mapzone/arena/share/ui/SharePanel.java
@@ -91,7 +91,7 @@ public class SharePanel
             shareContext.get().crs.set( mapViewer.maxExtent.get().getCoordinateReferenceSystem() );
         }
         if (!shareContext.get().resolution.isPresent()) {
-            shareContext.get().resolution.set( mapViewer.resolution.get() );
+            shareContext.get().resolution.set( (float)mapViewer.resolution.get() );
         }
         if (shareContext.get().selectionDescriptors.get().isEmpty()) {
             mapViewer.getLayers().stream()
